+++ build.rs
@@ -10,8 +10,8 @@ fn main() {
     let dest_file_name = Path::new(&dest_path).join("framework.min.js");
 
     // println!("Copy framework file {} to {}", source_path.to_str().unwrap(), dest_file_name.to_str().unwrap());
-    if !Path::exists(&dest_path) {
-        create_dir(&dest_path).unwrap();
+    if !Path::exists(dest_path) {
+        create_dir(dest_path).unwrap();
     }
     copy(&source_path, &dest_file_name).unwrap();
 }